pyproject.toml
[project]
name = "resumegenerator"
version = "0.1.0"
description = "CLI tool to parse LinkedIn PDF exports and generate polished CV documents"
requires-python = ">=3.13"
dependencies = [
    "anthropic",
    "pdfplumber",
    "pyyaml",
    "click",
    "weasyprint",
    "python-dotenv",
]

[project.scripts]
resume = "resume:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
